raw_syscalls: add lxc_raw_clone{_cb}()
