+++ src/tests/lxc_raw_clone.c
@@ -38,7 +38,7 @@
 #include <sys/types.h>
 
 #include "lxctest.h"
-#include "namespace.h"
+#include "raw_syscalls.h"
 #include "utils.h"
 
 int main(int argc, char *argv[])
